requirements.txt
argon2-cffi==21.1.0
asgiref==3.4.1
cffi==1.14.6
Django==3.2.6
pika==1.2.0
psycopg2==2.9.1
pycparser==2.20
PyJWT==2.1.0
python-decouple==3.4
pytz==2021.1
sqlparse==0.4.1
